Version 1.0
Working "Camera button" program. The c++ files have to be compiled(or the binaries can be used), and the autohotkey script has to be run.
If you want it to be active after every restart, you have to set the paths to the full paths of the binaries, and compile the AHK script, after which you have to place it to the Startup folder in Appdata